fix: Update color classes from 'text-primary-active' to 'text-primary' across various components for consistency
