+++ extensions/default/src/ViewerLayout/HeaderPatientInfo/HeaderPatientInfo.tsx
@@ -44,9 +44,9 @@ function HeaderPatientInfo({ servicesManager, appConfig }: withAppTypes) {
       onClick={handleOnClick}
     >
       {isMixedPatients ? (
-        <Icons.MultiplePatients className="text-primary-active" />
+        <Icons.MultiplePatients className="text-primary" />
       ) : (
-        <Icons.Patient className="text-primary-active" />
+        <Icons.Patient className="text-primary" />
       )}
       <div className="flex flex-col justify-center">
         {expanded ? (
@@ -61,12 +61,12 @@ function HeaderPatientInfo({ servicesManager, appConfig }: withAppTypes) {
             </div>
           </>
         ) : (
-          <div className="text-primary-active self-center text-[13px]">
+          <div className="text-primary self-center text-[13px]">
             {isMixedPatients ? 'Multiple Patients' : 'Patient'}
           </div>
         )}
       </div>
-      <Icons.ArrowLeft className={`text-primary-active ${expanded ? 'rotate-180' : ''}`} />
+      <Icons.ArrowLeft className={`text-primary ${expanded ? 'rotate-180' : ''}`} />
     </div>
   );
 }
