+++ extensions/default/src/utils/Toolbox.tsx
@@ -85,7 +85,7 @@ export function Toolbox({ buttonSectionId, title }: { buttonSectionId: string; t
         {CustomConfigComponent && (
           <div className="ml-auto mr-2">
             <Icons.Settings
-              className="text-primary-active h-4 w-4"
+              className="text-primary h-4 w-4"
               onClick={e => {
                 e.stopPropagation();
                 setShowConfig(!showConfig);
