+++ platform/ui-next/src/components/Header/Header.tsx
@@ -65,7 +65,7 @@ function Header({
             onClick={onClickReturn}
             data-cy="return-to-work-list"
           >
-            {isReturnEnabled && <Icons.ArrowLeft className="text-primary-active w-8" />}
+            {isReturnEnabled && <Icons.ArrowLeft className="text-primary w-8" />}
             <div className="ml-1">
               {WhiteLabeling?.createLogoComponentFn?.(React, props) || <Icons.OHIFLogo />}
             </div>
@@ -86,7 +86,7 @@ function Header({
                 <Button
                   variant="ghost"
                   size="icon"
-                  className="text-primary-active hover:bg-primary-dark mt-2 h-full w-full"
+                  className="text-primary hover:bg-primary-dark mt-2 h-full w-full"
                 >
                   <Icons.GearSettings />
                 </Button>
