+++ platform/ui-next/src/utils/getToggledClassName.tsx
@@ -1,6 +1,6 @@
 const getToggledClassName = isToggled => {
   return isToggled
-    ? '!text-primary-active'
+    ? '!text-primary'
     : '!text-common-bright hover:!bg-primary-dark hover:text-primary-light';
 };
